+++ src/components/index.js
@@ -1,0 +1,5 @@
+export {Navbar} from "./Navbar/Navbar";
+export {Sidebar} from "./Sidebar/Sidebar";
+export {VideoCard} from "./VideoCard/VideoCard";
+export {HistoryCard} from "./HistoryCard/HistoryCard";
+export {VideoIframe} from "./VideoIframe/VideoIframe";
