+++ src/pages/History/History.css
@@ -27,3 +27,10 @@
   font-weight: 700;
   cursor: pointer;
 }
+
+.videoCard-main-history-wrapper {
+  display: flex;
+  flex-direction: column;
+  padding: 0 18rem 5rem 5rem;
+  gap: 3rem;
+}
